psx-vector-api/pyproject.toml
[tool.poetry]
name = "psx-vector-api"
version = "0.1.0"
description = ""
authors = ["Willem Gillis", "Ismail Kutlu", "Maarten Taeymans"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.12"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
chromadb-client = "^0.5.4"
sentence_transformers = "^3.0.1"

[tool.poetry.dev-dependencies]

[tool.poetry.group.generators.dependencies]
datamodel-code-generator = "^0.25.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
